Moved backwards compatibility code from `v3` to `v2`
