+++ b2sdk/_v3/__init__.py
@@ -13,7 +13,7 @@
 
 # core
 
-from .api import B2Api
+from b2sdk.api import B2Api
 from b2sdk.api import Services
 from b2sdk.bucket import Bucket
 from b2sdk.bucket import BucketFactory
@@ -127,6 +127,7 @@ from b2sdk.stream import WritingStreamWithProgress
 
 from b2sdk.transfer.inbound.downloaded_file import DownloadedFile
 from b2sdk.transfer.inbound.downloaded_file import MtimeUpdatedFile
+from b2sdk.transfer.inbound.download_manager import DownloadManager
 
 from b2sdk.transfer.outbound.outbound_source import OutboundTransferSource
 from b2sdk.transfer.outbound.copy_source import CopySource
@@ -136,6 +137,7 @@ from b2sdk.transfer.outbound.upload_source import UploadSourceLocalFile
 from b2sdk.transfer.outbound.upload_source import UploadSourceLocalFileRange
 from b2sdk.transfer.outbound.upload_source import UploadSourceStream
 from b2sdk.transfer.outbound.upload_source import UploadSourceStreamRange
+from b2sdk.transfer.outbound.upload_manager import UploadManager
 
 from b2sdk.transfer.emerge.planner.upload_subpart import CachedBytesStreamOpener
 from b2sdk.transfer.emerge.write_intent import WriteIntent
